% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aml_pre}
\alias{aml_pre}
\title{AML_027: pre HSCT for AML}
\format{An object of class \code{dgTMatrix} with 12658 rows and 5150 columns.}
\usage{
aml_pre
}
\description{
The aml027 single-cell RNA-seq library was generated from cryopreserved bone
marrow mononuclear cell (BMMC) samples obtained from a patitent before
(aml027_pre) and after (aml027_post) undergoing hematopoietic stem cell
transplant (HSCT) for acute myeloid leukemia (AML).
}
\references{
Zheng, Grace XY, et al. "Massively parallel digital
    transcriptional profiling of single cells." Nature communications 8
    (2017): 14049.
}
\keyword{datasets}
