% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condcomp.R
\name{condcomp}
\alias{condcomp}
\title{Comparison of data conditions in a clustering}
\usage{
condcomp(clustering, cond, dmatrix, n = 1000, remove.na = TRUE)
}
\arguments{
\item{clustering}{A clustering of the data.}

\item{cond}{A factor indicating the condition which each data point is
subject to.}

\item{dmatrix}{A distance matrix describing the data to be analysed.}

\item{n}{The number of random silhouettes to be performed. Keep in mind that
the computation of several random silhouettes is the bottleneck of this
process.}

\item{remove.na}{Logical. Remove lines with NA (i.e. clusters which the
silhouette could not be computed).}
}
\value{
A data frame with various statistics regarding data heterogeneity
    inside each cluster.

Each row of the data frame contains several metrics regarding the conditions
found in an specific cluster. For now only a maximum of two conditions are
supported. These metrics are described below:

\describe{
\item{x_perc}{Numeric. The percentage of data points belonging to condition
'x'.}
\item{x_ratio}{Numeric. The ratio of data points belonging to condition 'x'.
For example, considering another condition 'y', the 'x_ratio' would be
computed as \eqn{x_perc / y_perc}.}
\item{true_sil}{Numeric. True silhouette. The silhouette for the data in
this cluster considering the conditions, as defined by the parameter
\code{cond}, as groups.}
\item{zscore}{Numeric. The Z-score computed based on the
\code{\link[cluster]{silhouette}}. See the 'Details' section.}
\item{pval}{Numeric. The p-value for 'true_sil'. Computed from the number of
Random Silhouettes (see 'Details') that are greater than the 'true_sil' for
this cluster.}
\item{iqr}{Factor. Interquartile Range (\link[outliers:scores]{IQR}) based
outlier detection. Considering the vector including the random silhouettes
(see 'Details') and the 'true_sil', the method checks whether 'true_sil' is
an outlier in said vector. This will be set to 'Diff' in case 'true_sil' is
an outlier or 'Same' otherwise.}
}
}
\description{
Performs a condition comparison on a given \code{clustering}. The comparison
is performed on each cluster separately between each condcition
(\code{cond}). Several statistics are used and, when analysed in conjunction,
they might give some insight regarding the heterogeneity of some of the
clusters.
}
\details{
For a given cluster, several metrics are computed, see the 'Return' section
for details about each metric. Some metrics make use of Random Silhouettes,
which is defined as follows: given a labeled data set, assign a random label
(from the set of labels) to each data point without changing the original
ratio of groups. Then compute the \code{\link{silhouette}} index for this
data considering these randomly assigned labels, the average silhouette width
is the Random Silhouette for the data (with randomly assigned labels). Being
an stochastic process, the Monte Carlo approach is applied which gives a
vector of several Random Silhouettes.
}
\examples{
clustering <- iris$Species
dmatrix <- as.matrix(dist(iris[-length(iris)]))
# Suppose the conditions are 'young' and 'old' fish
cond <- sample(c("young", "old"), nrow(iris), replace=TRUE)
comp <- condcomp(clustering, cond, dmatrix=dmatrix, n=10)
}
