% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condcomp.R
\name{condcompPlot}
\alias{condcompPlot}
\title{Plots the data frame of comparison of data conditions}
\usage{
condcompPlot(ccomp, col = ccomp$iqr, main = NULL, legend.title = "IQR")
}
\arguments{
\item{ccomp}{A data frame output from \code{\link{condcomp}}.}

\item{col}{Color parameter to be used. The default is to color according to
the IQR column of \code{ccomp}.}

\item{main}{Character vector (or expression) giving plot title.}

\item{legend.title}{Character vector giving the legend title.}
}
\value{
A ggplot2 object.
}
\description{
This function takes the output from \code{\link{condcomp}} and plots some of
its attributes in a scatter plot.
}
\details{
The first condition ratio that appears in the data frame will be plotted in
the y-axis (-log10 scale), whereas the Z-score will be plotted along the
x-axis. Each group will be colored by their respective IQR value as shown
in the legend.
}
\examples{
clustering <- iris$Species
dmatrix <- as.matrix(dist(iris[-length(iris)]))
# Suppose the conditions are 'young' and 'old' fish
cond <- sample(c("young", "old"), nrow(iris), replace=TRUE)
comp <- condcomp(clustering, cond, dmatrix=dmatrix, n=10)
condcompPlot(comp)
}
